+++ android/src/main/kotlin/com/cloudwebrtc/webrtc/FlutterWebRTCPlugin.kt
@@ -1,7 +1,6 @@
 package com.cloudwebrtc.webrtc
 
 import com.cloudwebrtc.webrtc.controller.MediaDevicesController
-import com.cloudwebrtc.webrtc.controller.OutputAudioDevicesController
 import com.cloudwebrtc.webrtc.controller.PeerConnectionFactoryController
 import com.cloudwebrtc.webrtc.controller.VideoRendererFactoryController
 import io.flutter.embedding.engine.plugins.FlutterPlugin
@@ -10,7 +9,6 @@ class FlutterWebRTCPlugin : FlutterPlugin {
     private var peerConnectionFactory: PeerConnectionFactoryController? = null
     private var mediaDevices: MediaDevicesController? = null
     private var videoRendererFactory: VideoRendererFactoryController? = null
-    private var outputAudioDevices: OutputAudioDevicesController? = null
 
     override fun onAttachedToEngine(registar: FlutterPlugin.FlutterPluginBinding) {
         val messenger = registar.binaryMessenger
@@ -20,7 +18,6 @@ class FlutterWebRTCPlugin : FlutterPlugin {
             PeerConnectionFactoryController(messenger, state)
         videoRendererFactory =
             VideoRendererFactoryController(messenger, registar.textureRegistry)
-        outputAudioDevices = OutputAudioDevicesController(messenger, state)
     }
 
     override fun onDetachedFromEngine(registrar: FlutterPlugin.FlutterPluginBinding) {}
