+++ lib/flutter_webrtc.dart
@@ -8,4 +8,3 @@ export 'src/platform/track.dart';
 export 'src/platform/native/video_view.dart'
     if (dart.library.html) 'src/universal/web/video_view.dart';
 export 'src/platform/video_renderer.dart';
-export 'src/platform/output_audio_devices.dart';
