+++ lib/src/api/devices.dart
@@ -54,3 +54,11 @@ Future<List<NativeMediaStreamTrack>> getDisplayMedia(
       .invokeMethod('getDisplayMedia', {'constraints': constraints.toMap()});
   return res.map((t) => NativeMediaStreamTrack.fromMap(t)).toList();
 }
+
+/// Switches output audio device to the provided [deviceId].
+///
+/// List of output audio devices can be obtained by [enumerateDevices].
+Future<void> setOutputAudioId(String deviceId) async {
+  await _mediaDevicesMethodChannel
+      .invokeMethod('setOutputAudioId', {'deviceId': deviceId});
+}
